requirements.txt
loguru~=0.7.2
art~=6.2
base58~=2.1.1
httpx~=0.27.0
mnemonic~=0.21
PyYAML~=6.0.1
undetected-playwright-patch